Rename kernel/tests
