+++ diffsky/experimental/kernels/tests/test_phot_kernels_merging.py
